requirements.txt
Django~=4.1
PyMySQL~=1.0.2
djangorestframework~=3.14.0
django-filter~=22.1